bring assertRegistrySubset() to registry_testing_util.go
