+++ helper/builtinplugins/registry_full_test.go
@@ -4,7 +4,6 @@
 package builtinplugins
 
 import (
-	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -26,20 +25,4 @@ func Test_newRegistry(t *testing.T) {
 
 	assertRegistrySubset(t, actual, expMinimal, "common")
 	assertRegistrySubset(t, actual, expFullAddon, "full addon")
-}
-
-func assertRegistrySubset(t *testing.T, r, subset *registry, subsetName string) {
-	t.Helper()
-
-	for k := range subset.credentialBackends {
-		require.Contains(t, r.credentialBackends, k, fmt.Sprintf("expected to contain %s auth backend", subsetName))
-	}
-
-	for k := range subset.databasePlugins {
-		require.Contains(t, r.databasePlugins, k, fmt.Sprintf("expected to contain %s database plugin", subsetName))
-	}
-
-	for k := range subset.logicalBackends {
-		require.Contains(t, r.logicalBackends, k, fmt.Sprintf("expected to contain %s logical backend", subsetName))
-	}
 }
